split extract object modules
