+++ autocti/__init__.py
@@ -26,15 +26,15 @@ from .charge_injection.imaging.settings import SettingsImagingCI
 from .charge_injection.imaging.simulator import SimulatorImagingCI
 from .charge_injection.layout import Layout2DCI
 from .cosmics.cosmics import SimulatorCosmicRayMap
-from .extract.two_d.parallel_overscan import Extract2DParallelOverscan
-from .extract.two_d.parallel_fpr import Extract2DParallelFPR
-from .extract.two_d.parallel_eper import Extract2DParallelEPER
-from .extract.two_d.serial_prescan import Extract2DSerialPrescan
-from .extract.two_d.serial_overscan import Extract2DSerialOverscan
-from .extract.two_d.serial_fpr import Extract2DSerialFPR
-from .extract.two_d.serial_eper import Extract2DSerialEPER
-from .extract.two_d.parallel_calibration import Extract2DParallelCalibration
-from .extract.two_d.serial_calibration import Extract2DSerialCalibration
+from .extract.two_d.parallel.overscan import Extract2DParallelOverscan
+from .extract.two_d.parallel.fpr import Extract2DParallelFPR
+from .extract.two_d.parallel.eper import Extract2DParallelEPER
+from .extract.two_d.serial.prescan import Extract2DSerialPrescan
+from .extract.two_d.serial.overscan import Extract2DSerialOverscan
+from .extract.two_d.serial.fpr import Extract2DSerialFPR
+from .extract.two_d.serial.eper import Extract2DSerialEPER
+from .extract.two_d.parallel.calibration import Extract2DParallelCalibration
+from .extract.two_d.serial.calibration import Extract2DSerialCalibration
 from .extract.two_d.master import Extract2DMaster
 from .charge_injection.fit import FitImagingCI
 from .mask.mask_2d import Mask2D
